// src/app/api/models/kinship.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Schema to hold Kinship information
 */
export interface Kinship {

  /**
   * Status
   */
  active?: boolean;

  /**
   * Kinship Code
   */
  code: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Kinship Name
   */
  name: string;
}